Added CreateGameToCriteriaTable

+++ server/libs/src/data_classes/board_data.go
@@ -10,6 +10,6 @@ type BoardData struct {
 	Anchor
 
 	Id, UserId, GameId, Name string
-	State, Active            int
+	State, Active            int64
 	DateAdded                time.Time
 }
